x #3522491, C# library: invalid size in Buffer_Continue_GoTo_Get, thanks to morpheusxx
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@4806 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfoDLL/MediaInfoDLL.cs
@@ -168,7 +168,7 @@ namespace MediaInfoLib
         }
         public Int64 Open_Buffer_Continue_GoTo_Get()
         {
-            return (int)MediaInfo_Open_Buffer_Continue_GoTo_Get(Handle);
+            return MediaInfo_Open_Buffer_Continue_GoTo_Get(Handle);
         }
         public int Open_Buffer_Finalize()
         {
